* The loop variable x was never used. Every iteration was re-checking the same Effect[0] slot. Spells with damage only on effect slot 1 or 2 were being falsely classified as having no damage component, incorrectly granting full mechanic immunity. (Effect[0] → Effect[x], EffectApplyAuraName[0] → EffectApplyAuraName[x]) * uint32 / uint32 integer division returns 0 for any health below max. Any m_FleeHealth > 0.0f (e.g., 0.2 for 20%) would evaluate as 0 < 0.2f = true, so every creature with flee configured fled on the first point of damage taken. (added (float) casts to both operands) * Both boundary guard checks in RemoveObject tested obj->GetPositionX() <= _minY — an X coordinate compared against a Y boundary. ChangeObjectLocation had the correct _minX at the same check. On asymmetric instance maps this would mis-classify valid X positions as out-of-bounds, skipping SetMapCell assignment or UpdateCellActivity and potentially causing null-cell dereferences later in the same function. * find() + conditional insert() or operator[] was doing two map lookups every call. operator[] on std::map performs insert-or-update in a single traversal. (SetWorldState collapsed to single operator[]) * find() returned and discarded, then operator[] did a second tree walk to return the value. Now the iterator from the first find() is used directly. (GetWorldState uses iterator from single find())

+++ Alpha/Core/src/ascent-world/AIInterface.cpp
@@ -1108,8 +1108,8 @@ void AIInterface::_UpdateCombat(uint32 p_time)
 			}
 			*/
 
-			if(m_canFlee && !m_hasFleed 
-				&& ((m_Unit->GetUInt32Value(UNIT_FIELD_HEALTH) / m_Unit->GetUInt32Value(UNIT_FIELD_MAXHEALTH)) < m_FleeHealth ))
+			if(m_canFlee && !m_hasFleed
+				&& ((float)m_Unit->GetUInt32Value(UNIT_FIELD_HEALTH) / (float)m_Unit->GetUInt32Value(UNIT_FIELD_MAXHEALTH)) < m_FleeHealth )
 				agent = AGENT_FLEE;
 			else if(m_canCallForHelp 
 				&& !m_hasCalledForHelp
